Обновление тестов на редьюсер, добавление теста на загрузку фильмов
