+++ package.json
@@ -14,6 +14,7 @@
     "@babel/core": "7.8.4",
     "@babel/preset-env": "7.8.4",
     "@babel/preset-react": "7.8.3",
+    "axios-mock-adapter": "1.18.1",
     "babel-jest": "25.1.0",
     "babel-loader": "8.0.6",
     "enzyme": "3.11.0",
